[REFACT] 작가가 자신의 프로그램을 생성 수정 조회 삭제하는 로직 Reservation 클래스로부터 분리
