+++ src/main/java/com/cuk/catsnap/domain/reservation/controller/PhotographerReservationController.java
@@ -1,13 +1,10 @@
 package com.cuk.catsnap.domain.reservation.controller;
 
 import com.cuk.catsnap.domain.reservation.dto.MonthReservationCheckListResponse;
-import com.cuk.catsnap.domain.reservation.dto.PhotographerProgramListResponse;
-import com.cuk.catsnap.domain.reservation.dto.photographer.request.ProgramRequest;
 import com.cuk.catsnap.domain.reservation.dto.photographer.request.ReservationTimeFormatRequest;
 import com.cuk.catsnap.domain.reservation.dto.photographer.response.PhotographerReservationInformationListResponse;
 import com.cuk.catsnap.domain.reservation.dto.photographer.response.ReservationTimeFormatIdResponse;
 import com.cuk.catsnap.domain.reservation.dto.photographer.response.ReservationTimeFormatListResponse;
-import com.cuk.catsnap.domain.reservation.dto.photographer.response.photographerProgramIdResponse;
 import com.cuk.catsnap.domain.reservation.entity.ReservationState;
 import com.cuk.catsnap.domain.reservation.entity.Weekday;
 import com.cuk.catsnap.domain.reservation.service.PhotographerReservationService;
@@ -199,62 +196,5 @@ public class PhotographerReservationController {
         photographerReservationService.unmappingWeekdayToReservationTimeFormatByWeekday(weekday);
         return ResultResponse.of(
             ReservationResultCode.PHOTOGRAPHER_RESERVATION_TIME_FORMAT_UNMAPPING_WEEKDAY);
-    }
-
-    @Operation(
-        summary = "작가가 자신의 예약 프로그램을 등록하거 수정(구현 완료)",
-        description = "작가가 자신의 예약 프로그램을 등록하거 수정하는 API입니다." +
-            "만약 수정을 원한다면, requestParameter로 programId 넘겨주어야 합니다." +
-            "내부적으로 프로그램 수정은 새로운 프로그램을 등록하기 때문에 프로그램 Id가 바뀜을 유의해야 합니다."
-    )
-    @ApiResponses({
-        @ApiResponse(responseCode = "201 SR010", description = "성공적으로가 예약 프로그램을 등록(수정)했습니다."),
-        @ApiResponse(responseCode = "404 SO000", description = "해당 게시물의 소유권을 찾을 수 없습니다.")
-    })
-    @PostMapping("/my/program")
-    public ResultResponse<photographerProgramIdResponse> postProgram(
-        @Parameter(description = "예약 프로그램", required = true)
-        @RequestBody
-        ProgramRequest programRequest,
-        @Parameter(description = "예약 프로그램의 id, 수정을 할 때만 입력", required = false)
-        @RequestParam(name = "programId", required = false)
-        Long programId
-    ) {
-        photographerProgramIdResponse photographerProgramIdResponse = photographerReservationService.createProgram(
-            programRequest, programId);
-        return ResultResponse.of(ReservationResultCode.PHOTOGRAPHER_POST_PROGRAM,
-            photographerProgramIdResponse);
-    }
-
-    @Operation(
-        summary = "작가가 자신의 예약 프로그램을 조회(구현 완료)",
-        description = "작가가 자신의 예약 프로그램을 조회하는 API입니다."
-    )
-    @ApiResponses(
-        @ApiResponse(responseCode = "200 SR013", description = "성공적으로가 예약 프로그램을 조회했습니다.")
-    )
-    @GetMapping("/my/program")
-    public ResultResponse<PhotographerProgramListResponse> getProgram() {
-        PhotographerProgramListResponse photographerProgramListResponse = photographerReservationService.getMyProgramList();
-        return ResultResponse.of(ReservationResultCode.PHOTOGRAPHER_LOOK_UP_PROGRAM,
-            photographerProgramListResponse);
-    }
-
-    @Operation(
-        summary = "작가가 자신의 예약 프로그램을 삭제(구현 완료)",
-        description = "작가가 자신의 예약 프로그램을 삭제하는 API입니다."
-    )
-    @ApiResponses({
-        @ApiResponse(responseCode = "200 SR011", description = "성공적으로가 예약 프로그램을 삭제했습니다."),
-        @ApiResponse(responseCode = "404 SO000", description = "해당 게시물의 소유권을 찾을 수 없습니다.")
-    })
-    @DeleteMapping("/my/program")
-    public ResultResponse<?> deleteProgram(
-        @Parameter(description = "삭제하고자 하는 예약 프로그램의 id", required = true)
-        @RequestParam("programId")
-        Long programId
-    ) {
-        photographerReservationService.softDeleteProgram(programId);
-        return ResultResponse.of(ReservationResultCode.PHOTOGRAPHER_DELETE_PROGRAM);
     }
 }
