+++ src/main/java/com/cuk/catsnap/domain/reservation/service/PhotographerReservationService.java
@@ -5,17 +5,12 @@ import com.cuk.catsnap.domain.photographer.repository.PhotographerRepository;
 import com.cuk.catsnap.domain.reservation.document.ReservationTimeFormat;
 import com.cuk.catsnap.domain.reservation.dto.MonthReservationCheckListResponse;
 import com.cuk.catsnap.domain.reservation.dto.MonthReservationCheckResponse;
-import com.cuk.catsnap.domain.reservation.dto.PhotographerProgramListResponse;
-import com.cuk.catsnap.domain.reservation.dto.PhotographerProgramResponse;
-import com.cuk.catsnap.domain.reservation.dto.photographer.request.ProgramRequest;
 import com.cuk.catsnap.domain.reservation.dto.photographer.request.ReservationTimeFormatRequest;
 import com.cuk.catsnap.domain.reservation.dto.photographer.response.PhotographerReservationInformationListResponse;
 import com.cuk.catsnap.domain.reservation.dto.photographer.response.PhotographerReservationInformationResponse;
 import com.cuk.catsnap.domain.reservation.dto.photographer.response.ReservationTimeFormatIdResponse;
 import com.cuk.catsnap.domain.reservation.dto.photographer.response.ReservationTimeFormatListResponse;
 import com.cuk.catsnap.domain.reservation.dto.photographer.response.ReservationTimeFormatResponse;
-import com.cuk.catsnap.domain.reservation.dto.photographer.response.photographerProgramIdResponse;
-import com.cuk.catsnap.domain.reservation.entity.Program;
 import com.cuk.catsnap.domain.reservation.entity.Reservation;
 import com.cuk.catsnap.domain.reservation.entity.ReservationState;
 import com.cuk.catsnap.domain.reservation.entity.Weekday;
@@ -137,48 +132,6 @@ public class PhotographerReservationService {
                 mapping.updateReservationTimeFormatId(null),
             () -> new OwnershipNotFoundException("내가 소유한 요일 중, 해당 요일을 찾을 수 없습니다.")
         );
-    }
-
-    /*
-     * Program은 수정을 하더라도 기존의 것을 soft delete하고 새로운 것을 생성하는 방식으로 진행한다.
-     * 왜냐하면 기존의 Program을 예약한 고객이 있을 수 있기 때문이다.
-     */
-    public photographerProgramIdResponse createProgram(
-        ProgramRequest programRequest, Long programId) {
-        Long photographerId = GetAuthenticationInfo.getUserId();
-        Photographer photographer = photographerRepository.getReferenceById(photographerId);
-        Program program = programRequest.toEntity(photographer);
-        if (programId != null) {
-            softDeleteProgram(programId);
-        }
-        Program savedProgram = programRepository.save(program);
-
-        return photographerProgramIdResponse.from(savedProgram);
-    }
-
-    public PhotographerProgramListResponse getMyProgramList() {
-        Long photographerId = GetAuthenticationInfo.getUserId();
-        List<Program> programList = programRepository.findByPhotographerIdAndDeletedFalse(
-            photographerId);
-        return PhotographerProgramListResponse.from(
-            programList.stream()
-                .map(PhotographerProgramResponse::from)
-                .toList()
-        );
-    }
-
-    public void softDeleteProgram(Long programId) {
-        Long photographerId = GetAuthenticationInfo.getUserId();
-        programRepository.findById(programId).ifPresentOrElse(
-            p -> {
-                if (p.getPhotographer().getId().equals(photographerId)) {
-                    p.softDelete();
-                } else {
-                    throw new OwnershipNotFoundException("내가 소유한 프로그램 중, 해당 프로그램을 찾을 수 없습니다.");
-                }
-            }, () -> {
-                throw new OwnershipNotFoundException("내가 소유한 프로그램 중, 해당 프로그램을 찾을 수 없습니다.");
-            });
     }
 
     public MonthReservationCheckListResponse getReservationListByMonth(LocalDate month) {
